+++ src/backends/scheduler/nameSearchHelpers.ts
